+++ src/helpers/index.ts
@@ -5,5 +5,6 @@ export {
   isChainSupported,
   doesChainSupportSplitManager,
   doesChainSupportMultiTokenFeeEscrow,
+  doesChainSupportPairedTokenLaunch,
 } from "./supportedChains";
 export * from "./permissions";